feat(src): add logic to change address/cancel order and updates throughout (workflow, activities, services, database crud)
The OrderWorkflow needs to be capable of cancelling an order and updating an address. Logically, both of these actions would occur before the shipping workflow. Therefore, the OrderWorkflow now stores an internal state to track these changes, updates this state in response to signals, and checks for state updates before every step in the workflow. If state changes, it calls relevant activities, which call signal service, which calls crud and updates database accordingly. Of relevance, compensatory actions are taken to refund customer if they have already paid.